Réorg de fichiers de code pour alléger database.js vers des fichiers plus spécialisés par domaine (dossier, groupe_instructeurs, personne)
